Refactor: Move the "Wallet Information" screen into a separate file
The method in main_window was getting a little long. Moved it to a
separate file to keep things sane(r).
